requirements.txt

torch~=1.13.1
transformers~=4.26.1
scipy~=1.10.1
pandas~=1.5.3
numpy~=1.24.2
tqdm~=4.65.0
scikit-learn~=1.2.2
Pillow
pydub~=0.25.1
xgboost
